% Generated by build script
\name{oecd_fetch}
\alias{oecd_fetch}
\title{Fetch OECD data using a popular dataflow shortname}
\usage{
oecd_fetch()
}
\arguments{
\item{name}{Short name: "NAAG", "CLI", "CPI", "ULC", "STLABOUR",}
\item{key}{Dimension key (default: all)}
\item{start_period}{Start year}
\item{end_period}{End year}
}
\value{tibble with SDMX columns}
\description{Fetch OECD data using a popular dataflow shortname}
